Replace `SyntaxRewriterFile.swift` with `SyntaxRewritersFile.swift`
